Clean /var/run before starting

+++ start.sh
@@ -9,8 +9,8 @@ chmod -R u+rw /log/forked-daapd /cache/forked-daapd
 sed -i -e "s/name = \"Music Server\"/name = \"${MUSIC_SERVER_NAME:-Music Server}\"/" /etc/forked-daapd.conf
 sed -i -e "s:/media:${MUSIC_SERVER_DATA:-/media}:" /etc/forked-daapd.conf
 
+[ -d "/var/run" ] && rm -rf /var/run/*
 mkdir -p /var/run/dbus
-rm -rf /var/run/*
 /usr/bin/dbus-daemon --system
 /usr/sbin/avahi-daemon --no-chroot -D
 /usr/sbin/forked-daapd -f -c /etc/forked-daapd.conf
